+++ morecantile/errors.py
@@ -15,3 +15,7 @@ class InvalidLatitudeError(MorecantileError):
 
 class TileArgParsingError(MorecantileError):
     """Raised when errors occur in parsing a function's tile arg(s)"""
+
+
+class PointOutsideTMSBounds(UserWarning):
+    """Point is outside TMS bounds."""